refactor(doctor): centralize count labels
