+++ src/commands/doctor-session-transcript-labels.ts
@@ -18,6 +18,7 @@ import {
 } from "../state/openclaw-agent-db.js";
 import { resolveTargetSqliteOptions } from "./doctor-session-sqlite-readers.js";
 import { ReadOnlySqliteTranscriptReader } from "./doctor-session-sqlite-transcript-readers.js";
+import { countLabel as formatCount } from "./doctor/shared/count-label.js";
 
 const NOTE_TITLE = "Session transcript labels";
 
@@ -180,10 +181,6 @@ function snapshotsMatch(
         row.seq === current[index]?.seq && row.eventJson === current[index]?.eventJson,
     )
   );
-}
-
-function formatCount(count: number, singular: string): string {
-  return `${count} ${singular}${count === 1 ? "" : "s"}`;
 }
 
 /** Reports or repairs legacy inbound-context labels in canonical SQLite transcripts. */
