+++ src/commands/doctor/shared/count-label.ts
